Rewrote Navigation
Changed how pages are loaded. Now the navigation header persistently exists and the body is instead changed by a javascript listener that will change the display based on the user clicking a link on the header. This removes the scripts for the other pages and instead just updates the existing webpage.